refactor: Move trace/logging classes into actually common library for sharing
